+++ CHANGELOG.md
@@ -9,8 +9,8 @@ API breaking changes
 - Multiple sources in `DownloadData`
 - Added `dirname` method in `SourceKind`
 - Verify file hashes in `DownloadData::download`
-- Added `rev` a user-provided file hash in `DownloadData`
-- Add rev and url source to profile imports
+- Added a user-provided file hash in `DownloadData`
+- Add file hash and url source to profile imports
 
 ## `ogj-v2.0.1-alpha`